split Main class into multiple files
