+++ src/main/scala/nl/knaw/dans/easy/stage/Constants.scala
@@ -1,0 +1,10 @@
+package nl.knaw.dans.easy.stage
+
+object Constants {
+  val DATASET_SDO       = "dataset"
+  val AMD_FILENAME      = "AMD"
+  val EMD_FILENAME      = "EMD"
+  val JSON_CFG_FILENAME = "cfg.json"
+  val FOXML_FILENAME    = "fo.xml"
+  val PRSQL_FILENAME    = "PRSQL"
+}
