+++ src/main/scala/nl/knaw/dans/easy/stage/Settings.scala
@@ -1,0 +1,5 @@
+package nl.knaw.dans.easy.stage
+
+import java.io.File
+
+case class Settings(ownerId: String, bagStorageLocation: String, bagitDir: File, sdoSetDir: File)
